+++ commons-lang/src/main/java/com/hb0730/commons/lang/bean/BeanException.java
@@ -1,0 +1,15 @@
+package com.hb0730.commons.lang.bean;
+
+import com.hb0730.commons.lang.exceptions.CommonsLangException;
+
+/**
+ * bean异常
+ *
+ * @author bing_huang
+ * @since 1.0.2
+ */
+public class BeanException extends CommonsLangException {
+    public BeanException(Throwable cause) {
+        super(cause);
+    }
+}