+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     + 新增ClassUtils中isNormalClass和isAbstract方法
     + 新增ReflectUtils中getField方法
     + 新增BeanInfoCache类
+    + 心中BeanUtils中对Bean描述的获取getPropertyDescriptor,getPropertyDescriptorMap等
 - commons-encrypt
     + 新增加密解密模块，对常规加密解密的支持MD5,SHA等
 ## Fixed